+++ CHANGELOG.md
@@ -5,8 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.3.2] - 2018-10-15
+### Fixed
+- phone number was mapped to the "company" field when adding/updating addresses
+
+### Changed
+- Shopify request log format changed to be compatible with `shopify-cart`'s log format
+
+## [1.3.1] - 2018-09-28
 ### Fixed
 - error handling for Shopify API and validation errors
+- customer not being redirected to app/web checkout after registration
+
+### Added
+- added customer address management synchronization
 
 ## [1.2.3] - 2018-08-30
 ### Fixed
@@ -20,7 +33,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 - user not logged in after registration in webcheckout
 
-[Unreleased]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Fheads%2Fmaster&sourceBranch=refs%2Ftags%2Fv1.2.2
+[Unreleased]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Fheads%2Fmaster&sourceBranch=refs%2Ftags%2Fv1.3.2
+[1.3.2]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.3.1&sourceBranch=refs%2Ftags%2Fv1.3.2
+[1.3.1]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.2.3&sourceBranch=refs%2Ftags%2Fv1.3.1
 [1.2.3]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.1.5&sourceBranch=refs%2Ftags%2Fv1.2.3
 [1.2.2]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.1.5&sourceBranch=refs%2Ftags%2Fv1.2.2
 [1.1.5]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/browse?at=refs%2Ftags%2Fv1.1.5